+++ Cargo.toml
@@ -1,7 +1,11 @@
 [workspace]
 members = [
     "axum",
+    "axum-core",
     "axum-debug",
     "axum-extra",
     "examples/*",
 ]
+
+[patch.crates-io]
+tower-http = { git = "https://github.com/tower-rs/tower-http.git" }
