+++ axum-core/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+categories = ["asynchronous", "network-programming", "web-programming"]
+description = "Core types and traits for axum"
+edition = "2018"
+homepage = "https://github.com/tokio-rs/axum"
+keywords = ["http", "web", "framework"]
+license = "MIT"
+name = "axum-core"
+readme = "README.md"
+repository = "https://github.com/tokio-rs/axum"
+version = "0.1.0"
+
+[dependencies]
+bytes = "1.0"
+http = "0.2"
+http-body = "0.4"
+hyper = { version = "0.14", default_features = false }
+
+[dev-dependencies]
+futures-util = "0.3"
+axum = { path = "../axum", version = "0.3" }
