+++ axum/Cargo.toml
@@ -20,6 +20,7 @@ tower-log = ["tower/log"]
 ws = ["tokio-tungstenite", "sha-1", "base64"]
 
 [dependencies]
+axum-core = { path = "../axum-core", version = "0.1" }
 async-trait = "0.1.43"
 bitflags = "1.0"
 bytes = "1.0"
