+++ axum/src/extract/matched_path.rs
@@ -114,7 +114,7 @@ mod tests {
                 .get::<MatchedPath>()
                 .unwrap()
                 .as_str()
-                .to_string();
+                .to_owned();
             req.extensions_mut().insert(MatchedPathFromMiddleware(path));
             self.0.call(req)
         }
@@ -127,7 +127,7 @@ mod tests {
     async fn access_matched_path() {
         let api = Router::new().route(
             "/users/:id",
-            get(|path: MatchedPath| async move { path.as_str().to_string() }),
+            get(|path: MatchedPath| async move { path.as_str().to_owned() }),
         );
 
         async fn handler(
@@ -146,7 +146,7 @@ mod tests {
         let app = Router::new()
             .route(
                 "/:key",
-                get(|path: MatchedPath| async move { path.as_str().to_string() }),
+                get(|path: MatchedPath| async move { path.as_str().to_owned() }),
             )
             .nest("/api", api)
             .nest(
