+++ axum/src/lib.rs
@@ -406,5 +406,5 @@ pub use self::json::Json;
 #[doc(inline)]
 pub use self::{error::Error, routing::Router};
 
-/// Alias for a type-erased error type.
-pub type BoxError = Box<dyn std::error::Error + Send + Sync>;
+#[doc(inline)]
+pub use axum_core::BoxError;
